+++ packages/components/atoms/f-link/src/components/_tests/Link.test.js
@@ -1,4 +1,4 @@
-import { shallowMount } from '@vue/test-utils';
+import { shallowMount, RouterLinkStub } from '@vue/test-utils';
 import VLink from '../Link.vue';
 
 const $style = {
@@ -8,10 +8,14 @@ const $style = {
     'o-link--noBreak': 'o-link--noBreak'
 };
 
+const mocks = {
+    $style
+};
+
 describe('Link', () => {
     it('should be defined', () => {
         // Arrange
-        const propsData = { };
+        const propsData = {};
 
         // Act
         const wrapper = shallowMount(VLink, {
@@ -22,20 +26,78 @@ describe('Link', () => {
         expect(wrapper.exists()).toBe(true);
     });
 
+    it('should render a router link when given the `to` attribute', () => {
+        // Arrange
+        const propsData = {
+            to: '/'
+        };
+
+        // Act
+        const wrapper = shallowMount(VLink, {
+            propsData,
+            mocks,
+            stubs: {
+                RouterLink: RouterLinkStub
+            }
+        });
+
+        const routerLink = wrapper.findComponent(RouterLinkStub);
+
+        // Assert
+        expect(routerLink.exists()).toBe(true);
+    });
+
+    it('should render an anchor tag when given the `href` attribute', () => {
+        // Arrange
+        const propsData = {
+            href: '/'
+        };
+
+        // Act
+        const wrapper = shallowMount(VLink, {
+            propsData,
+            mocks
+        });
+
+        const link = wrapper.find('[data-test-id="link-component"]');
+        const routerLink = wrapper.findComponent(RouterLinkStub);
+
+        // Assert
+        expect(link.element.tagName).toBe('A');
+        expect(routerLink.exists()).toBe(false);
+    });
+
+    it('should render an anchor tag when given both the `href` and `to` attributes', () => {
+        // Arrange
+        const propsData = {
+            href: '/',
+            to: '/'
+        };
+
+        // Act
+        const wrapper = shallowMount(VLink, {
+            propsData,
+            mocks
+        });
+
+        const link = wrapper.find('[data-test-id="link-component"]');
+        const routerLink = wrapper.findComponent(RouterLinkStub);
+
+        // Assert
+        expect(link.element.tagName).toBe('A');
+        expect(routerLink.exists()).toBe(false);
+    });
+
     describe('props :: ', () => {
         describe('isBold :: ', () => {
             it('should apply `o-link--bold` class to link if true', () => {
                 // Arrange
-                const propsData = {
-                    isBold: true
-                };
+                const propsData = { isBold: true };
 
                 // Act
                 const wrapper = shallowMount(VLink, {
                     propsData,
-                    mocks: {
-                        $style
-                    }
+                    mocks
                 });
 
                 const link = wrapper.find('[data-test-id="link-component"]');
@@ -48,16 +110,12 @@ describe('Link', () => {
         describe('hasTextDecoration :: ', () => {
             it('should apply `o-link--noDecoration` class to link if false', () => {
                 // Arrange
-                const propsData = {
-                    hasTextDecoration: false
-                };
+                const propsData = { hasTextDecoration: false };
 
                 // Act
                 const wrapper = shallowMount(VLink, {
                     propsData,
-                    mocks: {
-                        $style
-                    }
+                    mocks
                 });
 
                 const link = wrapper.find('[data-test-id="link-component"]');
@@ -70,16 +128,12 @@ describe('Link', () => {
         describe('isFullWidth :: ', () => {
             it('should apply `o-link--full` class to link if true', () => {
                 // Arrange
-                const propsData = {
-                    isFullWidth: true
-                };
+                const propsData = { isFullWidth: true };
 
                 // Act
                 const wrapper = shallowMount(VLink, {
                     propsData,
-                    mocks: {
-                        $style
-                    }
+                    mocks
                 });
 
                 const link = wrapper.find('[data-test-id="link-component"]');
@@ -92,16 +146,12 @@ describe('Link', () => {
         describe('noLineBreak :: ', () => {
             it('should apply `o-link--noBreak` class to link if true', () => {
                 // Arrange
-                const propsData = {
-                    noLineBreak: true
-                };
+                const propsData = { noLineBreak: true };
 
                 // Act
                 const wrapper = shallowMount(VLink, {
                     propsData,
-                    mocks: {
-                        $style
-                    }
+                    mocks
                 });
 
                 const link = wrapper.find('[data-test-id="link-component"]');
@@ -145,9 +195,7 @@ describe('Link', () => {
                     [externalLinkMessage, true]
                 ])('should return `%s` when `isExternalSite` is %s', (expected, isExternalSite) => {
                     // Arrange
-                    const propsData = {
-                        isExternalSite
-                    };
+                    const propsData = { isExternalSite };
 
                     // Act
                     const wrapper = shallowMount(VLink, {