+++ packages/components/atoms/f-link/stories/Link.stories.js
@@ -1,4 +1,4 @@
-import { boolean } from '@storybook/addon-knobs';
+import { boolean, text } from '@storybook/addon-knobs';
 import { withA11y } from '@storybook/addon-a11y';
 import VLink from '../src/components/Link.vue';
 
@@ -15,6 +15,18 @@ export const VLinkComponent = () => ({
         };
     },
     props: {
+        href: {
+            default: text('Anchor link path', '/')
+        },
+
+        to: {
+            default: text('Router link path (href must be empty for this to be applied)', '')
+        },
+
+        isRouterLink: {
+            default: boolean('Is a router link?', false)
+        },
+
         isExternalSite: {
             default: boolean('Opens an external site', false)
         },
@@ -51,7 +63,10 @@ export const VLinkComponent = () => ({
 
     template: `<v-link
                     :data-test-id="dataTestId"
-                    href="https://www.just-eat.co.uk/"
+                    v-bind="{
+                        ...(href ? { href } : to ? { to } : {})
+                    }"
+                    :is-router-link="isRouterLink"
                     :is-bold="isBold"
                     :has-text-decoration="hasTextDecoration"
                     :is-full-width="isFullWidth"