+++ packages/components/molecules/f-restaurant-card/CHANGELOG.md
@@ -3,6 +3,36 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+v0.5.0
+------------------------------
+*October 28, 2021*
+
+### Added
+- Basic dish item styling
+### Updated
+- Broke down parts of card into subcomponents
+- Add hover styling to card title
+- Remove some placeholder text
+- Move logo out of image container for potential fallback options in future
+
+v0.4.0
+------------------------------
+*October 25, 2021*
+
+### Added
+- Injectable error boundary to wrap around slots and subcomponents
+
+v0.3.0
+------------------------------
+*October 22, 2021*
+
+### Added
+- Basic component styling
+
+### Updated
+- Component HTML to use new classes, attributes and fewer elements
+- Storybook args updated to use new component props contract
+
 v0.2.0
 ------------------------------
 *October 7, 2021*